+++ docs/release-notes.md
@@ -9,13 +9,13 @@
 
 ## Upcoming release
 * For players:
+  * Added support for unofficial 64-bit Stardew Valley, including automatic support in the SMAPI installer.
+  * Added update checks for Stardew64Installer if it patched the game.
   * When many mods fail to load, root dependencies are now listed in their own group so it's easier to see which ones you should try updating first.
   * On macOS, the `StardewModdingAPI.bin.osx` file is no longer overwritten if it's identical to avoid resetting file permissions (thanks to 007wayne!).
-  * Added update checks for Stardew64Installer if it patched the game.
   * Fixed error for non-English players after returning to title, reloading, and entering town with a completed movie theater.
   * Fixed `world_clear` console command not removing resource clumps outside the farm and secret woods.
   * Fixed inconsistent spelling/style for 'macOS'.
-  * Internal changes to prepare for unofficial 64-bit.
 
 * For modders:
   * Added asset propagation for `Data\Concessions`.